+++ src/pdf.rs
@@ -6,7 +6,7 @@ use crate::utilities::math::fmax;
 use crate::utilities::onb::ONB;
 use crate::Vector3;
 
-const PI: f64 = std::f64::consts::PI;
+const PI: f32 = std::f32::consts::PI;
 pub enum PDFType<'a> {
     PDFObj { pdf: PDF<'a> },
     PDFCosine { pdf: PDFCosine },
@@ -15,16 +15,16 @@ pub enum PDFType<'a> {
 }
 
 impl PDFType<'_> {
-    pub fn value(&self, direction: Vector3<f64>) -> f64 {
+    pub fn value(&self, direction: Vector3<f32>) -> f32 {
         match self {
             Self::PDFObj { pdf } => {
-                let acc: f64 = pdf
+                let acc: f32 = pdf
                     .objects
                     .iter()
                     .map(|object| object.pdf_value(pdf.o, direction))
                     .sum();
 
-                acc / pdf.objects.len() as f64
+                acc / pdf.objects.len() as f32
             }
             Self::PDFCosine { pdf } => {
                 let cosine = Vector3::dot(direction.norm(), pdf.onb.w);
@@ -42,7 +42,7 @@ impl PDFType<'_> {
         }
     }
 
-    pub fn generate(&self, rng: &mut ThreadRng) -> Vector3<f64> {
+    pub fn generate(&self, rng: &mut ThreadRng) -> Vector3<f32> {
         match self {
             Self::PDFObj { pdf } => pdf.objects.choose(rng).unwrap().random(pdf.o, rng),
             Self::PDFCosine { pdf } => pdf.onb.local(Vector3::random_cosine_direction(rng)),
@@ -62,12 +62,12 @@ impl PDFType<'_> {
 
 #[allow(clippy::upper_case_acronyms)]
 pub struct PDF<'a> {
-    o: Vector3<f64>,
+    o: Vector3<f32>,
     objects: &'a [Object],
 }
 
 impl<'a> PDF<'a> {
-    pub fn new(o: Vector3<f64>, objects: &'a [Object]) -> Self {
+    pub fn new(o: Vector3<f32>, objects: &'a [Object]) -> Self {
         Self { o, objects }
     }
 }
@@ -77,7 +77,7 @@ pub struct PDFCosine {
 }
 
 impl PDFCosine {
-    pub fn new(w: Vector3<f64>) -> Self {
+    pub fn new(w: Vector3<f32>) -> Self {
         Self {
             onb: ONB::build_from(w),
         }
@@ -93,14 +93,14 @@ impl PDFSphere {
 }
 
 pub struct PDFBlinnPhongSpec {
-    r_in_direction: Vector3<f64>,
+    r_in_direction: Vector3<f32>,
     onb: ONB,
-    normal: Vector3<f64>,
-    exponent: f64,
+    normal: Vector3<f32>,
+    exponent: f32,
 }
 
 impl PDFBlinnPhongSpec {
-    pub fn new(r_in_direction: Vector3<f64>, normal: Vector3<f64>, exponent: f64) -> Self {
+    pub fn new(r_in_direction: Vector3<f32>, normal: Vector3<f32>, exponent: f32) -> Self {
         let reflected = Vector3::reflect(r_in_direction.norm(), normal);
 
         let onb = ONB::build_from(reflected);
@@ -123,12 +123,12 @@ impl<'a> PDFMixture<'a> {
     pub fn new(p: &'a PDFType, q: &'a PDFType) -> Self {
         Self { p, q }
     }
-    pub fn value(&self, direction: Vector3<f64>) -> f64 {
+    pub fn value(&self, direction: Vector3<f32>) -> f32 {
         0.5 * self.p.value(direction) + 0.5 * self.q.value(direction)
     }
 
-    pub fn generate(&self, rng: &mut ThreadRng) -> Vector3<f64> {
-        if rng.gen::<f64>() < 0.5 {
+    pub fn generate(&self, rng: &mut ThreadRng) -> Vector3<f32> {
+        if rng.gen::<f32>() < 0.5 {
             self.p.generate(rng)
         } else {
             self.q.generate(rng)
