+++ src/sphere.rs
@@ -9,13 +9,13 @@ use crate::{
 };
 #[derive(Clone)]
 pub struct Sphere {
-    pub center: Vector3<f64>,
-    pub radius: f64,
+    pub center: Vector3<f32>,
+    pub radius: f32,
     pub material: Material,
     bounding_box: AABB,
 }
 impl Sphere {
-    pub fn new(center: Vector3<f64>, radius: f64, material: Material) -> Self {
+    pub fn new(center: Vector3<f32>, radius: f32, material: Material) -> Self {
         let radius_v = Vector3::new(radius, radius, radius);
         let bounding_box = AABB::new(center - radius_v, center + radius_v);
         Self {
@@ -26,8 +26,8 @@ impl Sphere {
         }
     }
 
-    pub fn get_sphere_uv(p: Vector3<f64>) -> (f64, f64) {
-        let pi = std::f64::consts::PI;
+    pub fn get_sphere_uv(p: Vector3<f32>) -> (f32, f32) {
+        let pi = std::f32::consts::PI;
         let theta = (-p.y).acos();
         let phi = (-p.z).atan2(p.x) + pi;
 
@@ -36,22 +36,22 @@ impl Sphere {
         (u, v)
     }
 
-    pub fn pdf_value(&self, origin: Vector3<f64>, v: Vector3<f64>) -> f64 {
-        if let Some(_hit) = self.hit(&Ray::new(origin, v), 0.001, f64::MAX) {
+    pub fn pdf_value(&self, origin: Vector3<f32>, v: Vector3<f32>) -> f32 {
+        if let Some(_hit) = self.hit(&Ray::new(origin, v), 0.001, f32::MAX) {
             let cos_theta_max =
                 (1.0 - self.radius * self.radius / (self.center - origin).magnitude2()).sqrt();
-            let solid_angle = 2.0 * std::f64::consts::PI * (1.0 - cos_theta_max);
+            let solid_angle = 2.0 * std::f32::consts::PI * (1.0 - cos_theta_max);
             return 1.0 / solid_angle;
         }
         0.0
     }
 
-    pub fn random(&self, origin: Vector3<f64>, rng: &mut ThreadRng) -> Vector3<f64> {
-        pub fn random_to_sphere(radius: f64, distance_2: f64, rng: &mut ThreadRng) -> Vector3<f64> {
-            let r1 = rng.gen::<f64>();
-            let r2 = rng.gen::<f64>();
+    pub fn random(&self, origin: Vector3<f32>, rng: &mut ThreadRng) -> Vector3<f32> {
+        pub fn random_to_sphere(radius: f32, distance_2: f32, rng: &mut ThreadRng) -> Vector3<f32> {
+            let r1 = rng.gen::<f32>();
+            let r2 = rng.gen::<f32>();
             let z = 1.0 + r2 * ((1.0 - radius * radius / distance_2).sqrt() - 1.0);
-            let phi = 2.0 * std::f64::consts::PI * r1;
+            let phi = 2.0 * std::f32::consts::PI * r1;
             let x = phi.cos() * (1.0 - z * z).sqrt();
             let y = phi.sin() * (1.0 - z * z).sqrt();
             Vector3::new(x, y, z)
@@ -64,7 +64,7 @@ impl Sphere {
 }
 impl Hittable for Sphere {
     #[inline(always)]
-    fn hit(&self, r: &crate::ray::Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, r: &crate::ray::Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let a = r.direction.magnitude2();
         let oc = r.origin - self.center;
         let c = oc.magnitude2() - self.radius * self.radius;
