+++ src/transformations.rs
@@ -8,12 +8,12 @@ use crate::{
 #[derive(Clone)]
 pub struct Translate {
     object: Box<Object>,
-    offset: Vector3<f64>,
+    offset: Vector3<f32>,
     bounding_box: AABB,
 }
 
 impl Translate {
-    pub fn new(object: Object, offset: Vector3<f64>) -> Self {
+    pub fn new(object: Object, offset: Vector3<f32>) -> Self {
         let bb_object = object.bounding_box();
         let bounding_box = AABB::new(bb_object.minimum + offset, bb_object.maximum + offset);
 
@@ -26,7 +26,7 @@ impl Translate {
 }
 
 impl Hittable for Translate {
-    fn hit(&self, r: &crate::ray::Ray, t_min: f64, t_max: f64) -> Option<crate::ray::HitRecord> {
+    fn hit(&self, r: &crate::ray::Ray, t_min: f32, t_max: f32) -> Option<crate::ray::HitRecord> {
         let moved_r = Ray::new(r.origin - self.offset, r.direction);
         if let Some(mut hit) = self.object.hit(&moved_r, t_min, t_max) {
             hit.p += self.offset;
@@ -43,29 +43,29 @@ impl Hittable for Translate {
 #[derive(Clone)]
 pub struct RotateY {
     object: Box<Object>,
-    sin_theta: f64,
-    cos_theta: f64,
+    sin_theta: f32,
+    cos_theta: f32,
     bounding_box: AABB,
 }
 
 impl RotateY {
-    pub fn new(object: Object, angle: f64) -> Self {
+    pub fn new(object: Object, angle: f32) -> Self {
         let radians = angle.to_radians();
         let sin_theta = radians.sin();
         let cos_theta = radians.cos();
 
         let bb = object.bounding_box();
 
-        let mut min_acc = Vector3::new(f64::MAX, f64::MAX, f64::MAX);
-        let mut max_acc = Vector3::new(f64::MIN, f64::MIN, f64::MIN);
+        let mut min_acc = Vector3::new(f32::MAX, f32::MAX, f32::MAX);
+        let mut max_acc = Vector3::new(f32::MIN, f32::MIN, f32::MIN);
 
         for i in 0..2 {
             for j in 0..2 {
                 for k in 0..2 {
                     let vec = Vector3::new(
-                        i as f64 * bb.maximum.x + (1 - i) as f64 * bb.minimum.x,
-                        j as f64 * bb.maximum.y + (1 - j) as f64 * bb.minimum.y,
-                        k as f64 * bb.maximum.z + (1 - k) as f64 * bb.minimum.z,
+                        i as f32 * bb.maximum.x + (1 - i) as f32 * bb.minimum.x,
+                        j as f32 * bb.maximum.y + (1 - j) as f32 * bb.minimum.y,
+                        k as f32 * bb.maximum.z + (1 - k) as f32 * bb.minimum.z,
                     );
 
                     let new_vec = rot(vec, sin_theta, cos_theta);
@@ -86,7 +86,7 @@ impl RotateY {
 }
 
 impl Hittable for RotateY {
-    fn hit(&self, r: &crate::ray::Ray, t_min: f64, t_max: f64) -> Option<crate::ray::HitRecord> {
+    fn hit(&self, r: &crate::ray::Ray, t_min: f32, t_max: f32) -> Option<crate::ray::HitRecord> {
         let origin = rot(r.origin, -self.sin_theta, self.cos_theta);
         let direction = rot(r.direction, -self.sin_theta, self.cos_theta);
         let rotated_r = Ray::new(origin, direction);
@@ -105,7 +105,7 @@ impl Hittable for RotateY {
     }
 }
 
-fn rot(p: Vector3<f64>, sin_theta: f64, cos_theta: f64) -> Vector3<f64> {
+fn rot(p: Vector3<f32>, sin_theta: f32, cos_theta: f32) -> Vector3<f32> {
     Vector3::new(
         Vector3::dot(p, Vector3::new(cos_theta, 0., sin_theta)),
         Vector3::dot(p, Vector3::new(0., 1., 0.)),
